+++ .env.example
@@ -1,5 +1,5 @@
-MONGODB_ROOT_USERNAME=always-on-root
-MONGODB_ROOT_PASSWORD=always-on-root-pw
+MONGODB_ROOT_USERNAME=testerozza-root
+MONGODB_ROOT_PASSWORD=testerozza-root-pw
 
-DB_USERNAME=always-on
-DB_PASSWORD=always-on-pw
+DB_USERNAME=ttestofferson
+DB_PASSWORD=P@ssw0rd1!
