+++ README.md
@@ -1,32 +1,32 @@
-# AlwaysON
+# Testerozza
 
-  AlwaysON is the API that always responds with the EXACT response you want.
+  Testerozza is the API that always responds with the EXACT response you want.
 
-**NOTE: As of right now AlwaysON only supports the HTTP protocol but eventually we would like it to support additional protocols such as gRPC, GraphQL etc.**
+**NOTE: As of right now Testerozza only supports the HTTP protocol but eventually we would like it to support additional protocols such as gRPC, GraphQL etc.**
 
 
-If you're looking to contribute to AlwaysON please read the [Contributing Guidelines](./CONTRIBUTING.md) which will help you oriented with the project, what
+If you're looking to contribute to Testerozza please read the [Contributing Guidelines](./CONTRIBUTING.md) which will help you oriented with the project, what
 we expect from contributors and how to setup your development environment.
 
 ## Terminology
 
-- `Request`: An Http request received by AlwaysON.
-- `Response`: An Http response that AlwaysON has been configured to use as a
+- `Request`: An Http request received by Testerozza.
+- `Response`: An Http response that Testerozza has been configured to use as a
               response to a specific Request.
-- `Echo Endpoint`: A REST endpoint in AlwaysON that imitates another API. When a
-                   Request is recieved AlwaysON will respond with the configured
+- `Echo Endpoint`: A REST endpoint in Testerozza that imitates another API. When a
+                   Request is recieved Testerozza will respond with the configured
                    Response for that specific Request.
-- `System Endpoint`: A REST Endpoint used to configure, and manage how AlwaysON
+- `System Endpoint`: A REST Endpoint used to configure, and manage how Testerozza
                      will behave when imitating an API.
 - `Manifest`: A JSON object (not necessarily a file but it could be) thats maps
               a single Request to a single Response or series of Responses that
-              AlwaysON returns when the Request is received by an echo endpoint.
+              Testerozza returns when the Request is received by an echo endpoint.
 
 
 ## Introduction
- - What is AlwaysON
- - Why did I build AlwaysON
- - What problem(s) is AlwaysON solving
+ - What is Testerozza
+ - Why did I build Testerozza
+ - What problem(s) is Testerozza solving
 
 
 ## Future Development
