+++ testerozza/controllers/status.py
@@ -3,8 +3,8 @@ from typing import Annotated, Dict
 from fastapi import APIRouter, Depends
 from fastapi.responses import JSONResponse
 
-from alwayson.config import Config
-from alwayson.db.client import DbClient
+from testerozza.config import Config
+from testerozza.db.client import DbClient
 
 router = APIRouter()
