+++ tests/test.env
@@ -1,4 +1,4 @@
-DB_USERNAME=always-on
-DB_PASSWORD=always-on-pw
+DB_USERNAME=testerozza
+DB_PASSWORD=testerozza-pw
 DB_HOST=localhost
 DB_PORT=27017
